refactor: Extract Click-dependent functions to output.py, remove duplicate is_full_url
Architecture improvements:
- Move TextStyle, fail_out, failures_out, print2 from utility.py to
  new utility/output.py module (removes Click dependency from utility)
- Remove duplicate is_full_url() from cli_utility.py (was marked as
  DUPLICATE with TODO to remove), re-export from utility instead
- Remove unused urllib3.parse_url import from cli_utility.py
- Update __init__.py to re-export from both modules for backward compat
