+++ yojenkins/utility/__init__.py
@@ -1,3 +1,4 @@
 """Importing sub-modules"""
 
+from .output import TextStyle, fail_out, failures_out, print2
 from .utility import *
